Moved test files for modules into subfolder
